Separation of headers ans sources files
