+++ Headers/constants.h
@@ -1,0 +1,6 @@
+#pragma once
+
+// Constants
+const auto MINIMAL_PRICE{0};
+const auto MAXIMAL_PRICE{10'000};
+const auto MAXIMAL_PRICE_EASY{1'000};
